+++ src/accessory/onoff/onoff.ts
@@ -4,12 +4,12 @@ import { DummyAccessory, DummyAccessoryDependency } from '../base.js';
 
 import { strings } from '../../i18n/i18n.js';
 
-import { HKCharacteristicKey, isValidOnState, OnState, printableValues } from '../../model/enums.js';
+import { HKCharacteristicKey, OnState, SensorBehavior } from '../../model/enums.js';
 import { HistoryType } from '../../model/history.js';
 import { OnOffConfig } from '../../model/types.js';
 import { Values, Webhook } from '../../model/webhook.js';
 
-import { storageGet_Deprecated } from '../../tools/storage.js';
+import { isValid, printableValues } from '../../tools/validation.js';
 
 export abstract class OnOffAccessory<C extends OnOffConfig = OnOffConfig> extends DummyAccessory<C> {
 
@@ -18,7 +18,7 @@ export abstract class OnOffAccessory<C extends OnOffConfig = OnOffConfig> extend
   constructor(dependency: DummyAccessoryDependency<C>) {
     super(dependency);
 
-    if (!isValidOnState(this.config.defaultState)) {
+    if (!isValid(OnState, this.config.defaultState)) {
       this.log.warning(strings.onOff.badDefault, this.name, `'${dependency.config.defaultState}'`, printableValues(OnState));
     }
 
@@ -33,7 +33,7 @@ export abstract class OnOffAccessory<C extends OnOffConfig = OnOffConfig> extend
 
   override get webhooks(): Webhook[] {
     return [
-      new Webhook(this.identifier, HKCharacteristicKey.On,
+      new Webhook(this, HKCharacteristicKey.On,
         new Values( [true, false], 'true, false'),
         () => this.on,
         (value, syncOnly) => {
@@ -54,7 +54,7 @@ export abstract class OnOffAccessory<C extends OnOffConfig = OnOffConfig> extend
       return;
     }
 
-    const on = this.getProperty(HKCharacteristicKey.On) ?? await storageGet_Deprecated(`${this.identifier}:DefaultState`);
+    const on = this.getProperty(HKCharacteristicKey.On);
     if (on === undefined) {
       await this.registerStateChange();
       return;
@@ -109,9 +109,9 @@ export abstract class OnOffAccessory<C extends OnOffConfig = OnOffConfig> extend
     this.service.updateCharacteristic(this.Characteristic.On, this.on);
 
     if (this.sensor) {
-      if (!this.sensor.timerControlled) {
+      if (this.sensor.behavior === SensorBehavior.MIRROR) {
         this.sensor.active = this.on !== this.defaultState;
-      } else if (this.on !== this.defaultState) {
+      } else if (this.sensor.behavior === SensorBehavior.TIMER && this.on !== this.defaultState) {
         this.sensor.active = false;
       }
     }
@@ -126,7 +126,7 @@ export abstract class OnOffAccessory<C extends OnOffConfig = OnOffConfig> extend
   override async reset(): Promise<void> {
     if (this.on !== this.defaultState) {
       await this.setOn(this.defaultState);
-      if (this.sensor?.timerControlled) {
+      if (this.sensor?.behavior === SensorBehavior.TIMER) {
         this.sensor.active = true;
       }
     }