+++ src/homebridge/platform.ts
@@ -14,7 +14,6 @@ import { History } from '../model/history.js';
 import { DummyConfig, DummyPlatformConfig, GroupConfig } from '../model/types.js';
 import { WebhookManager } from '../model/webhook.js';
 
-import { migrateAccessories } from '../tools/configMigration.js';
 import { Log } from '../tools/log.js';
 import { Storage } from '../tools/storage.js';
 import getVersion from '../tools/version.js';
@@ -37,14 +36,13 @@ export class HomebridgeDummyPlatform implements DynamicPlatformPlugin {
     private readonly api: API,
   ) {
 
-    const userLang = Intl.DateTimeFormat().resolvedOptions().locale.split('-')[0];
-    setLanguage(userLang);
+    setLanguage(api.user.configPath());
 
     this.Service = api.hap.Service;
     this.Characteristic = api.hap.Characteristic;
 
     this.log = new Log(logger, config.verbose === true);
-    this.webhookManager = new WebhookManager(this.log, config.webhookPort);
+    this.webhookManager = new WebhookManager(this.log, this.api.user.configPath(), { port: config.webhookPort, ...config.webhookConfig });
     this.conditionManager = new ConditionManager(this.log, api.user.storagePath());
 
     this.log.always(
@@ -76,6 +74,7 @@ export class HomebridgeDummyPlatform implements DynamicPlatformPlugin {
     this.dummyAccessories.forEach( accessory => {
       accessory.teardown();
     });
+    this.webhookManager.teardown();
     this.conditionManager.teardown();
   }
 
@@ -86,11 +85,6 @@ export class HomebridgeDummyPlatform implements DynamicPlatformPlugin {
     const keepIdentifiers = new Set<string>();
 
     const accessories: DummyConfig[] = this.config.accessories || [];
-    if (this.config.migrationNeeded) {
-      const migratedAccessories = await migrateAccessories(this.log, this.api.user.configPath()) ?? [];
-      accessories.push(...migratedAccessories);
-    }
-
     const groupAccessories = new Map<string, GroupConfig>();
 
     const history = new History(this.api, this.log);
@@ -129,7 +123,7 @@ export class HomebridgeDummyPlatform implements DynamicPlatformPlugin {
         continue;
       }
 
-      if (accessoryConfig.enableWebook) {
+      if (accessoryConfig.enableWebhook === true || accessoryConfig.enableWebook === true) {
         this.webhookManager.registerWebhooks(dummyAccessory.webhooks);
       }