+++ src/i18n/missing/el.ts
@@ -1,0 +1,24 @@
+const el = {
+
+  config: {
+
+    enumNames: {
+      sensorMirror: 'Mirror Accessory',
+      sensorTimer: 'Timer Controlled',
+    },
+
+    title: {
+      sensorBehavior: 'Sensor Behavior',
+    },
+
+  },
+
+  sensor: {
+
+    badBehavior: '%s has invalid sensor behavior %s. Must be one of: %s', // accessory name, input, list of type names
+
+  },
+
+};
+
+export default el;