+++ src/model/webhook.ts
@@ -1,8 +1,14 @@
 import express, { Request, Response } from 'express';
+import { readFileSync } from 'fs';
 import { CharacteristicValue } from 'homebridge';
 import { Server } from 'http';
+import { createServer, ServerOptions } from 'https';
+import path from 'path';
+
+import { DummyAccessory } from '../accessory/base.js';
 
 import { HKCharacteristicKey } from './enums.js';
+import { DummyConfig, WebhookConfig } from './types.js';
 
 import { strings } from '../i18n/i18n.js';
 
@@ -12,7 +18,7 @@ import { assert } from '../tools/validation.js';
 
 const DEFAULT_PORT = 63743;
 
-type WebhookGetter = () => (CharacteristicValue);
+type WebhookGetter = () => (CharacteristicValue | undefined);
 type WebhookSetter = (value: CharacteristicValue, syncOnly: boolean) => (string);
 
 export class Range {
@@ -26,7 +32,7 @@ export class Values {
 export class Webhook {
 
   constructor(
-    public readonly id: string,
+    public readonly accessory: DummyAccessory<DummyConfig>,
     public readonly characteristic: HKCharacteristicKey,
     public readonly validValues: Range | Values,
     public readonly getter: WebhookGetter,
@@ -43,14 +49,41 @@ export class WebhookManager {
 
   constructor(
     private readonly log: Log,
-    private readonly port: number | undefined,
+    configPath: string,
+    private readonly config: WebhookConfig = {},
   ) {
 
-    this.port = port ?? DEFAULT_PORT;
+    try {
+      const systemConfig = readFileSync(configPath, { encoding: 'utf8' });
+      const systemSSLConfig = JSON.parse(systemConfig).platforms.filter( (c: Record<string, string>) => c.platform === 'config')[0].ssl;
 
-    if (typeof this.port !== 'number') {
+      if (systemSSLConfig !== undefined) {
+
+        if (systemSSLConfig.selfSigned === true) {
+          systemSSLConfig.key = systemSSLConfig.key ?? path.join(configPath, '../ssl-certs/private-key.pem');
+          systemSSLConfig.cert = systemSSLConfig.cert ?? path.join(configPath, '../ssl-certs/certificate.pem');
+        }
+
+        this.config = { ...systemSSLConfig , ...this.config };
+      }
+
+    } catch {
+      // Nothing
+    }
+
+    this.config.port = this.config.port ?? DEFAULT_PORT;
+
+    if (typeof this.config.port !== 'number') {
       log.error(strings.webhook.badPort, DEFAULT_PORT);
-      this.port = DEFAULT_PORT;
+      this.config.port = DEFAULT_PORT;
+    }
+
+    for (const keyString of ['key', 'cert', 'pfx', 'passphrase']) {
+      const key = keyString as keyof WebhookConfig;
+      if (this.config[key] !== undefined && typeof this.config[key] !== 'string') {
+        log.error(strings.webhook.badSSLParameter, `'${key}'`, '\'string\'', `'${typeof this.config[key]}'`);
+        this.config[key] = undefined;
+      }
     }
   }
 
@@ -58,7 +91,7 @@ export class WebhookManager {
     for (const webhook of webhooks) {
       this.webhooks.push(webhook);
       if (!webhook.disableLogging) {
-        this.log.always(strings.webhook.register, `\`${webhook.id}\`` , `\`${webhook.characteristic}\``);
+        this.log.always(strings.webhook.register, webhook.accessory.name, `\`${webhook.accessory.identifier}\`` , `\`${webhook.characteristic}\``);
       }
     }
   }
@@ -85,8 +118,34 @@ export class WebhookManager {
       this.onRequest(request, response);
     });
 
-    this.server = exp.listen(this.port, () => {
-      this.log.always(strings.webhook.started, this.port);
+    if (this.config.disableSSL !== true) {
+
+      try {
+
+        let credentials: ServerOptions | undefined;
+        if (this.config.pfx !== undefined) {
+          credentials = { pfx: readFileSync(this.config.pfx), passphrase: this.config.passphrase };
+        } else if (this.config.key !== undefined && this.config.cert !== undefined) {
+          credentials = { key: readFileSync(this.config.key), cert: readFileSync(this.config.cert) };
+        }
+
+        if (credentials !== undefined) {
+
+          this.server = createServer(credentials, exp).listen(this.config.port, () => {
+            this.log.always(`${strings.webhook.started} (https)`, this.config.port);
+          });
+
+          return;
+        }
+
+      } catch (err) {
+        this.log.error(strings.webhook.badSSL);
+      }
+
+    }
+
+    this.server = exp.listen(this.config.port, () => {
+      this.log.always(`${strings.webhook.started} (http)`, this.config.port);
     });
   }
 
@@ -100,6 +159,12 @@ export class WebhookManager {
   private onRequest(request: Request, response: Response) {
 
     const data = { ...request.query, ...request.body };
+
+    if (Object.keys(data).length === 0) {
+      this.onList(response);
+      return;
+    }
+
     this.log.ifVerbose(`${strings.webhook.received}\n${JSON.stringify(data)}`);
 
     if (!assert(this.log, 'Webhook', data, 'id')) {
@@ -140,6 +205,11 @@ export class WebhookManager {
     response.status(200).json({ value: value });
   }
 
+  private onList(response: Response) {
+    const html = this.generateTableHTML();
+    response.status(200).contentType('html').send(html);
+  }
+
   private setValue(
     response: Response, id: string, characteristic: HKCharacteristicKey, value: CharacteristicValue, syncOnly: boolean) {
 
@@ -170,7 +240,7 @@ export class WebhookManager {
 
   private getWebhook(response: Response, id: string, characteristic: HKCharacteristicKey): Webhook | undefined {
 
-    const byId = this.webhooks.filter( (webhook) => webhook.id === id);
+    const byId = this.webhooks.filter( (webhook) => webhook.accessory.identifier === id);
     if (byId.length === 0) {
       this.onBadRequest(response, strings.webhook.unregisteredId.replace('%s', `\`${id}\``));
       return;
@@ -196,5 +266,64 @@ export class WebhookManager {
     if(alsoLog) {
       this.log.error(errorMessage);
     }
+  }
+
+  private generateTableHTML(): string {
+
+    return  `
+<!DOCTYPE html>
+
+<html>
+
+<head>
+
+<title>${strings.webhook.title}</title>
+
+<style>
+table {
+  font-family: arial, sans-serif;
+  border-collapse: collapse;
+}
+
+td, th {
+  border: 1px solid #bbbbbb;
+  text-align: left;
+  padding: 8px 32px 8px 16px;
+}
+
+tr:nth-child(even) {
+  background-color: #dddddd;
+}
+</style>
+
+</head>
+<body>
+
+<h2>${strings.webhook.title}</h2>
+<table>
+  <tr>
+    <th>${strings.config.title.name}</th>
+    <th>${strings.accessory.identifier}</th>
+    <th>${strings.webhook.command}</th>
+    <th>${strings.webhook.values}</th>
+    <th>${strings.webhook.example}</th>
+  </tr>
+  ${this.webhooks.map(webhook => {
+    const exampleValue = webhook.validValues instanceof Range ? webhook.validValues.min : webhook.validValues.values[0];
+    return `
+  <tr>
+    <td>${webhook.accessory.name}</td>
+    <td>${webhook.accessory.identifier}</td>
+    <td>${webhook.characteristic}</td>
+    <td>${webhook.validValues instanceof Range ? `${webhook.validValues.min} - ${webhook.validValues.max}` : webhook.validValues.asString}</td>
+    <td><a target="_blank" href=?id=${webhook.accessory.identifier}&command=${webhook.characteristic}&value=${exampleValue}>${strings.webhook.link}</td>
+  </tr>
+  `;
+  }).join('')}
+</table>
+
+</body>
+</html>
+`;
   }
 }